chore(jazz-run): chore extract the accountCreate command
